wip more tests moving around
