+++ core/src/test/java/overflowdb/testdomains/gratefuldead/Artist.java
