+++ tinkerpop3/src/test/java/overflowdb/TraversalOptimizationTest.java
@@ -4,6 +4,7 @@ import org.junit.Test;
 import overflowdb.testdomains.gratefuldead.Artist;
 import overflowdb.testdomains.gratefuldead.FollowedBy;
 import overflowdb.testdomains.gratefuldead.GratefulDead;
+import overflowdb.testdomains.gratefuldead.GratefulDeadTp3;
 import overflowdb.testdomains.gratefuldead.Song;
 import overflowdb.testdomains.gratefuldead.SungBy;
 import overflowdb.testdomains.gratefuldead.WrittenBy;
@@ -17,7 +18,7 @@ public class TraversalOptimizationTest {
 
   @Test
   public void optimizationStrategyAffectedSteps() throws IOException {
-    try (OdbGraphTp3 graph = GratefulDead.openAndLoadSampleData()) {
+    try (OdbGraphTp3 graph = GratefulDeadTp3.openAndLoadSampleData()) {
       // using `g.V().hasLabel(lbl)` optimization
       assertEquals(584, graph.traversal().V().hasLabel(Song.label).toList().size());
       assertEquals(142, graph.traversal().V().has(Song.PERFORMANCES, 1).toList().size());
