+++ tinkerpop3/src/test/java/overflowdb/testdomains/gratefuldead/GratefulDeadTp3.java
@@ -8,21 +8,10 @@ import overflowdb.tinkerpop.OdbGraphTp3;
 import java.io.IOException;
 import java.util.Arrays;
 
-public class GratefulDead {
-  public static Graph open() {
-    return open(Config.withoutOverflow());
-  }
-
-  public static Graph open(Config config) {
-    return Graph.open(
-        config,
-        Arrays.asList(Song.factory, Artist.factory),
-        Arrays.asList(FollowedBy.factory, SungBy.factory, WrittenBy.factory)
-    );
-  }
+public class GratefulDeadTp3 {
 
   public static OdbGraphTp3 openAndLoadSampleData() throws IOException {
-    OdbGraphTp3 graph = OdbGraphTp3.wrap(open());
+    OdbGraphTp3 graph = OdbGraphTp3.wrap(GratefulDead.open());
     loadData(graph);
     return graph;
   }
@@ -30,7 +19,7 @@ public class GratefulDead {
   public static OdbGraphTp3 openAndLoadSampleData(String path) throws IOException {
     OdbGraphTp3 graph =
       OdbGraphTp3.wrap(
-        open(Config.withDefaults().withStorageLocation(path))
+        GratefulDead.open(Config.withDefaults().withStorageLocation(path))
       );
     loadData(graph);
     return graph;
